Moved generic calls to fram.c/.h
