+++ stm32/lib/fram/src/fm24cl16b.c
@@ -105,22 +105,6 @@ FramStatus Fm24cl16bRead(uint16_t addr, uint8_t len, uint8_t *data) {
   return FRAM_OK;
 }
 
-HAL_StatusTypeDef ConfigureSettings(configuration c) {
-  HAL_StatusTypeDef status = HAL_OK;
-
-  // TODO(GSOC student) implement user config write
-
-  return status;
-}
-
-configuration ReadSettings(void) {
-  configuration c;
-
-  // TODO(GSOC student) implement user config read
-
-  return c;
-}
-
 FramAddress FramConvertAddrMem(uint16_t addr) {
   // convert flat address space to matrix
   FramAddress addr_mat;
