+++ stm32/lib/fram/src/fram.c
@@ -1,0 +1,17 @@
+#include "fram.h"
+
+HAL_StatusTypeDef ConfigureSettings(configuration c) {
+  HAL_StatusTypeDef status = HAL_OK;
+
+  // TODO(GSOC student) implement user config write
+
+  return status;
+}
+
+configuration ReadSettings(void) {
+  configuration c;
+
+  // TODO(GSOC student) implement user config read
+
+  return c;
+}
